Move session config to common
